firefox: Fix <browser> setup to specify remoteness.
This is needed so that the URL is loaded out of process properly.

Otherwise it'll try to load it in process, and fail due to security
checks.

The nice thing is that now the web application runs properly sandboxed
and that a bug in it, or in the engine, can't own your computer.